+++ packages/expo-modules-autolinking/src/index.ts
@@ -4,6 +4,7 @@ import path from 'path';
 import { patchReactImportsAsync } from './ReactImportsPatcher';
 import {
   findModulesAsync,
+  generateModulesProviderAsync,
   generatePackageListAsync,
   getProjectPackageJsonPathAsync,
   mergeLinkingOptionsAsync,
@@ -160,7 +161,7 @@ module.exports = async function (args: string[]) {
   }).option<boolean>('-j, --json', 'Output results in the plain JSON format.', () => true, false);
 
   // Generates a source file listing all packages to link.
-  // It's deprecated, use `generate-modules-provider` instead.
+  // It's deprecated for apple platforms, use `generate-modules-provider` instead.
   registerResolveCommand<GenerateOptions>('generate-package-list', async (results, options) => {
     const modules = options.empty ? [] : await resolveModulesAsync(results, options);
     generatePackageListAsync(modules, options);
@@ -187,13 +188,14 @@ module.exports = async function (args: string[]) {
       const modules = await resolveModulesAsync(results, options);
       const filteredModules = modules.filter((module) => packages.includes(module.packageName));
 
-      generatePackageListAsync(filteredModules, options);
+      generateModulesProviderAsync(filteredModules, options);
     }
   )
     .option(
       '-t, --target <path>',
       'Path to the target file, where the package list should be written to.'
     )
+    .option('--entitlement <path>', 'Path to the Apple code signing entitlements file.')
     .option(
       '-p, --packages <packages...>',
       'Names of the packages to include in the generated modules provider.'