+++ packages/expo-router/src/__tests__/navigation.test.ios.tsx
@@ -11,9 +11,47 @@ import {
   usePathname,
   Link,
 } from '../exports';
+import { Drawer } from '../layouts/Drawer';
 import { Stack } from '../layouts/Stack';
 import { Tabs } from '../layouts/Tabs';
 import { act, fireEvent, renderRouter, screen } from '../testing-library';
+
+it('should respect `unstable_settings', () => {
+  const render = (options: any = {}) =>
+    renderRouter(
+      {
+        '(one,two)/_layout': {
+          unstable_settings: {
+            initialRouteName: 'apple',
+            two: {
+              initialRouteName: 'orange',
+            },
+          },
+          default: () => <Tabs />,
+        },
+        '(one,two)/apple': () => <Text testID="apple"> Apple</Text>,
+        '(two)/banana': () => <Text testID="banana">Banana</Text>,
+        '(one,two)/orange': () => <Text testID="orange">Orange</Text>,
+      },
+      options
+    );
+
+  render({ initialUrl: '/orange' });
+  expect(screen).toHaveSegments(['(two)', 'orange']);
+
+  expect(screen.getByTestId('orange')).toBeVisible();
+  // Orange is the initial route so you can't go back
+  expect(router.canGoBack()).toBeFalsy();
+
+  // Reset the app, but start at /banana
+  screen.unmount();
+  render({ initialUrl: '/banana' });
+
+  expect(screen.getByTestId('banana')).toBeVisible();
+  // Orange should be the initialRouteName, because we are in (two)
+  act(() => router.back());
+  expect(screen.getByTestId('orange')).toBeVisible();
+});
 
 describe('hooks only', () => {
   it('can handle navigation between routes', async () => {
@@ -1470,4 +1508,71 @@ it('respects nested unstable settings', async () => {
   expect(screen.getByTestId('profile')).toBeVisible();
   fireEvent.press(screen.getByText('Home'));
   expect(screen.getByTestId('index')).toBeVisible();
+});
+
+describe('navigation action fallbacks', () => {
+  function runPushTest() {
+    act(() => router.navigate('/'));
+    expect(screen).toHavePathname('/');
+
+    // Go to one
+    act(() => router.navigate('/one'));
+    expect(screen).toHavePathname('/one');
+    expect(screen.getByTestId('one')).toBeOnTheScreen();
+
+    // Push to two. `PUSH` action should fall back to `NAVIGATE` action
+    act(() => router.push('/two'));
+    expect(screen).toHavePathname('/two');
+    expect(screen.getByTestId('two')).toBeOnTheScreen();
+  }
+
+  function runReplaceTest() {
+    act(() => router.navigate('/'));
+    expect(screen).toHavePathname('/');
+
+    // Go to one
+    act(() => router.navigate('/one'));
+    expect(screen).toHavePathname('/one');
+    expect(screen.getByTestId('one')).toBeOnTheScreen();
+
+    // Replace to two. `REPLACE` action should fall back to `JUMP_TO` action
+    act(() => router.replace('/two'));
+    expect(screen).toHavePathname('/two');
+    expect(screen.getByTestId('two')).toBeOnTheScreen();
+  }
+
+  function runRedirectionTest() {
+    act(() => router.navigate('/'));
+    expect(screen).toHavePathname('/');
+
+    // `<Redirect />` uses `REPLACE` action and should fall back to `JUMP_TO` action
+    act(() => router.navigate('/redirected'));
+    expect(screen).toHavePathname('/');
+  }
+
+  it('can fall back correctly for tab navigators', () => {
+    renderRouter({
+      _layout: () => <Tabs />,
+      one: () => <Text testID="one" />,
+      two: () => <Text testID="two" />,
+      redirected: () => <Redirect href="/" />,
+    });
+
+    runPushTest();
+    runReplaceTest();
+    runRedirectionTest();
+  });
+
+  it('can fall back correctly for drawer navigators', () => {
+    renderRouter({
+      _layout: () => <Drawer useLegacyImplementation={false} />,
+      one: () => <Text testID="one" />,
+      two: () => <Text testID="two" />,
+      redirected: () => <Redirect href="/" />,
+    });
+
+    runPushTest();
+    runReplaceTest();
+    runRedirectionTest();
+  });
 });